Added user reviews.
Users can now choose a park then press the compose button to create a review. The user enters ratings for the five categories then can submit the review. The firebase database is updated and the user is taken back to the park detail view and the averages are updated on that view as well.