Added Firebase Database to project.
When user views a park, the app checks to see if park is in the database and if it isn't then adds the park into the database. User's can input ratings on parks and add a comment then the ratings and comment is saved into the database and average ratings are displayed on the park detail screen.